Move NamedplacesGuesser to a separate file #1989
